requirements.txt
annotated-types==0.7.0
anyio==4.6.0
APScheduler==3.10.4
click==8.1.7
exceptiongroup==1.2.2
fastapi==0.115.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
psycopg2==2.9.9
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
pytz==2024.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
SQLAlchemy-Utils==0.41.2
starlette==0.38.6
typing_extensions==4.12.2
tzlocal==5.2
uvicorn==0.31.0